php E-commerce Website
E-commerce web application built using php routing. Instead of relying on the webserver to map the request path to a file, all requests are forwarded to index.php which has defined routes and callbacks registered to each route. If the request URI is a valid route, the callback returns a page to the user else, redirected to the 404 page